# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "## TMF API Reference: TMF633 - Service Catalog Management\n\n### Release\
    \ : 18.5 - December 2018\n\nService Catalog API is one of Catalog Management API\
    \ Family. Service Catalog API goal is to provide a catalog of services. \n\n###\
    \ Resource \n- serviceCatalog\n\n### Operations\nService Catalog API performs\
    \ the following operations on the resource :\n- Retrieve an entity or a collection\
    \ of entities depending on filter criteria\n- Partial update of an entity (including\
    \ updating rules)\n- Create an entity (including default values and creation rules)\n\
    - Delete an entity (for administration purposes)\n- Manage notification of events"
  version: "4.0.0"
  title: "Service Catalog Management"
host: "serverRoot"
basePath: "/tmf-api/serviceCatalogManagement/v4/"
...
schemes:
- "https"
consumes:
- "application/json;charset=utf-8"
produces:
- "application/json;charset=utf-8"
paths:
...
  /serviceSpecification:
    get:
      tags:
      - "serviceSpecification"
      summary: "List or find ServiceSpecification objects"
      description: "This operation list or find ServiceSpecification entities"
      operationId: "listServiceSpecification"
      parameters:
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to be provided in response"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "Requested index for start of resources to be provided in response"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Requested number of resources to be provided in response"
        required: false
        type: "integer"
      responses:
        200:
          description: "Success"
          headers:
            X-Total-Count:
              type: "integer"
              description: "Total number of items matching criteria"
            X-Result-Count:
              type: "integer"
              description: "Actual number of items returned in the response body"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ServiceSpecification"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceSpecification"
    post:
      tags:
      - "serviceSpecification"
      summary: "Creates a ServiceSpecification"
      description: "This operation creates a ServiceSpecification entity."
      operationId: "createServiceSpecification"
      parameters:
      - in: "body"
        name: "serviceSpecification"
        description: "The ServiceSpecification to be created"
        required: true
        schema:
          $ref: "#/definitions/ServiceSpecification_Create"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/ServiceSpecification"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceSpecification"
  /serviceSpecification/{id}:
    get:
      tags:
      - "serviceSpecification"
      summary: "Retrieves a ServiceSpecification by ID"
      description: "This operation retrieves a ServiceSpecification entity. Attribute\
        \ selection is enabled for all first level attributes."
      operationId: "retrieveServiceSpecification"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the ServiceSpecification"
        required: true
        type: "string"
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to provide in response"
        required: false
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/ServiceSpecification"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceSpecification"
    delete:
      tags:
      - "serviceSpecification"
      summary: "Deletes a ServiceSpecification"
      description: "This operation deletes a ServiceSpecification entity."
      operationId: "deleteServiceSpecification"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the ServiceSpecification"
        required: true
        type: "string"
      responses:
        204:
          description: "Deleted"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceSpecification"
    patch:
      tags:
      - "serviceSpecification"
      summary: "Updates partially a ServiceSpecification"
      description: "This operation updates partially a ServiceSpecification entity."
      operationId: "patchServiceSpecification"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the ServiceSpecification"
        required: true
        type: "string"
      - in: "body"
        name: "serviceSpecification"
        description: "The ServiceSpecification to be updated"
        required: true
        schema:
          $ref: "#/definitions/ServiceSpecification_Update"
      responses:
        200:
          description: "Updated"
          schema:
            $ref: "#/definitions/ServiceSpecification"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceSpecification"
...
definitions:
...
  RelatedParty:
    type: "object"
    properties:
      id:
        type: "string"
        description: "Unique identifier of a related party"
      href:
        type: "string"
        description: "Reference of the related party, could be a party reference or\
          \ a party role reference"
      name:
        type: "string"
        description: "Name of the related party"
      role:
        type: "string"
        description: "Role of the related party."
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
      '@referredType':
        type: "string"
        description: "The actual type of the target instance when needed for disambiguation."
    description: "RelatedParty reference. A related party defines party or party role\
      \ linked to a specific entity."
    example:
      '@referredType': "@referredType"
      role: "role"
      '@baseType': "@baseType"
      '@type': "@type"
      name: "name"
      id: "id"
      href: "href"
      '@schemaLocation': "@schemaLocation"
...
  ServiceSpecification:
    type: "object"
    properties:
      id:
        type: "string"
        description: "Unique identifier of the service specification"
      href:
        type: "string"
        description: "Reference of the service specification"
      description:
        type: "string"
        description: "A narrative that explains in detail what the service specification\
          \ is"
      isBundle:
        type: "boolean"
        description: "isBundle determines whether a ServiceSpecification represents\
          \ a single ServiceSpecification (false), or a bundle of ServiceSpecification\
          \ (true)."
      lastUpdate:
        type: "string"
        format: "date-time"
        description: "Date and time of the last update of the service specification"
      lifecycleStatus:
        type: "string"
        description: "Used to indicate the current lifecycle status of the service\
          \ specification"
      name:
        type: "string"
        description: "Name of the service specification"
      version:
        type: "string"
        description: "Service specification version"
      attachment:
        type: "array"
        description: "A list of attachments (Attachment [*]). Complements the description\
          \ of the specification through video, pictures..."
        items:
          $ref: "#/definitions/AttachmentRef"
      relatedParty:
        type: "array"
        description: "A list of related party references (RelatedParty [*]). A related\
          \ party defines party or party role linked to a specific entity."
        items:
          $ref: "#/definitions/RelatedParty"
      resourceSpecification:
        type: "array"
        description: "A list of resource specification references (ResourceSpecificationRef\
          \ [*]). The ResourceSpecification is required for a service specification\
          \ with type ResourceFacingServiceSpecification (RFSS)."
        items:
          $ref: "#/definitions/ResourceSpecificationRef"
      serviceLevelSpecification:
        type: "array"
        description: "A list of service level specifications related to this service\
          \ specification, and which will need to be satisifiable for corresponding\
          \ service instances; e.g. Gold, Platinum"
        items:
          $ref: "#/definitions/ServiceLevelSpecificationRef"
      serviceSpecCharacteristic:
        type: "array"
        description: "A list of service spec characteristics (ServiceSpecCharacteristic\
          \ [*]). This class represents the key features of this service specification."
        items:
          $ref: "#/definitions/ServiceSpecCharacteristic"
      serviceSpecRelationship:
        type: "array"
        description: "A list of service specifications related to this specification,\
          \ e.g. migration, substitution, dependency or exclusivity relationship"
        items:
          $ref: "#/definitions/ServiceSpecRelationship"
      targetServiceSchema:
        description: "A target service schema reference (TargetServiceSchemaRef).\
          \ The reference object to the schema and type of target service which is\
          \ described by service specification."
        $ref: "#/definitions/TargetServiceSchema"
      validFor:
        description: "The period for which the service specification is valid"
        $ref: "#/definitions/TimePeriod"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
    description: "ServiceSpecification is a class that offers characteristics to describe\
      \ a type of service.\nFunctionally, it acts as a template by which Services\
      \ may be instantiated. By sharing the same  specification, these services would\
      \ therefore share the same set of characteristics."
    example:
      isBundle: true
      targetServiceSchema:
        '@baseType': "@baseType"
        '@type': "@type"
        '@schemaLocation': "@schemaLocation"
      lifecycleStatus: "lifecycleStatus"
      serviceLevelSpecification:
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
      validFor:
        startDateTime: "2000-01-23T04:56:07.000+00:00"
        endDateTime: "2000-01-23T04:56:07.000+00:00"
      '@type': "@type"
      resourceSpecification:
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
        version: "version"
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
        version: "version"
      description: "description"
      relatedParty:
      - '@referredType': "@referredType"
        role: "role"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "@schemaLocation"
      - '@referredType': "@referredType"
        role: "role"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "@schemaLocation"
      version: "version"
      attachment:
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        description: "description"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
        url: "url"
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        description: "description"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
        url: "url"
      '@baseType': "@baseType"
      lastUpdate: "2000-01-23T04:56:07.000+00:00"
      name: "name"
      id: "id"
      href: "href"
      '@schemaLocation': "http://example.com/aeiou"
      serviceSpecCharacteristic:
      - validFor:
          startDateTime: "2000-01-23T04:56:07.000+00:00"
          endDateTime: "2000-01-23T04:56:07.000+00:00"
        '@type': "@type"
        '@valueSchemaLocation': "@valueSchemaLocation"
        isUnique: true
        description: "description"
        serviceSpecCharacteristicValue:
        - rangeInterval: "rangeInterval"
          isDefault: true
          regex: "regex"
          valueTo: 5
          unitOfMeasure: "unitOfMeasure"
          '@baseType': "@baseType"
          validFor:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          '@type': "@type"
          valueType: "valueType"
          '@schemaLocation': "http://example.com/aeiou"
          value: ""
          valueFrom: 1
        - rangeInterval: "rangeInterval"
          isDefault: true
          regex: "regex"
          valueTo: 5
          unitOfMeasure: "unitOfMeasure"
          '@baseType': "@baseType"
          validFor:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          '@type': "@type"
          valueType: "valueType"
          '@schemaLocation': "http://example.com/aeiou"
          value: ""
          valueFrom: 1
        maxCardinality: 0
        minCardinality: 6
        serviceSpecCharRelationship:
        - role: "role"
          relationshipType: "relationshipType"
          '@baseType': "@baseType"
          validFor:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          '@type': "@type"
          name: "name"
          id: "id"
          href: "href"
          '@schemaLocation': "http://example.com/aeiou"
        - role: "role"
          relationshipType: "relationshipType"
          '@baseType': "@baseType"
          validFor:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          '@type': "@type"
          name: "name"
          id: "id"
          href: "href"
          '@schemaLocation': "http://example.com/aeiou"
        regex: "regex"
        '@baseType': "@baseType"
        valueType: "valueType"
        name: "name"
        '@schemaLocation': "http://example.com/aeiou"
        extensible: true
        configurable: true
      - validFor:
          startDateTime: "2000-01-23T04:56:07.000+00:00"
          endDateTime: "2000-01-23T04:56:07.000+00:00"
        '@type': "@type"
        '@valueSchemaLocation': "@valueSchemaLocation"
        isUnique: true
        description: "description"
        serviceSpecCharacteristicValue:
        - rangeInterval: "rangeInterval"
          isDefault: true
          regex: "regex"
          valueTo: 5
          unitOfMeasure: "unitOfMeasure"
          '@baseType': "@baseType"
          validFor:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          '@type': "@type"
          valueType: "valueType"
          '@schemaLocation': "http://example.com/aeiou"
          value: ""
          valueFrom: 1
        - rangeInterval: "rangeInterval"
          isDefault: true
          regex: "regex"
          valueTo: 5
          unitOfMeasure: "unitOfMeasure"
          '@baseType': "@baseType"
          validFor:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          '@type': "@type"
          valueType: "valueType"
          '@schemaLocation': "http://example.com/aeiou"
          value: ""
          valueFrom: 1
        maxCardinality: 0
        minCardinality: 6
        serviceSpecCharRelationship:
        - role: "role"
          relationshipType: "relationshipType"
          '@baseType': "@baseType"
          validFor:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          '@type': "@type"
          name: "name"
          id: "id"
          href: "href"
          '@schemaLocation': "http://example.com/aeiou"
        - role: "role"
          relationshipType: "relationshipType"
          '@baseType': "@baseType"
          validFor:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          '@type': "@type"
          name: "name"
          id: "id"
          href: "href"
          '@schemaLocation': "http://example.com/aeiou"
        regex: "regex"
        '@baseType': "@baseType"
        valueType: "valueType"
        name: "name"
        '@schemaLocation': "http://example.com/aeiou"
        extensible: true
        configurable: true
      serviceSpecRelationship:
      - role: "role"
        relationshipType: "relationshipType"
        '@baseType': "@baseType"
        validFor:
          startDateTime: "2000-01-23T04:56:07.000+00:00"
          endDateTime: "2000-01-23T04:56:07.000+00:00"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
      - role: "role"
        relationshipType: "relationshipType"
        '@baseType': "@baseType"
        validFor:
          startDateTime: "2000-01-23T04:56:07.000+00:00"
          endDateTime: "2000-01-23T04:56:07.000+00:00"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
...
  Error:
    type: "object"
    required:
    - "code"
    - "reason"
    properties:
      code:
        type: "integer"
        description: "Application relevant detail, defined in the API or a common\
          \ list."
      reason:
        type: "string"
        description: "Explanation of the reason for the error which can be shown to\
          \ a client user."
      message:
        type: "string"
        description: "More details and corrective actions related to the error which\
          \ can be shown to a client user."
      status:
        type: "integer"
        description: "HTTP Error code extension"
      referenceError:
        type: "string"
        format: "uri"
        description: "URI of documentation describing the error."
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class."
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name."
    description: "Used when an API throws an Error, typically with a HTTP error response-code\
      \ (3xx, 4xx, 5xx)"